HackInt/requirements.txt
django
djangorestframework
faker
pillow
djangorestframework-simplejwt
